+++ src/device/stub/device_pelion_time.c
@@ -13,10 +13,10 @@
  *
  *
  * This method causes the current-thread to sleep for <b>us</b>
- * milli-seconds.
+ * micro-seconds.
  */
 void
-pelion_delay_us(unsigned int us) {
+pelion_delay_us(unsigned long long us) {
 
 }
 
@@ -26,7 +26,7 @@ pelion_delay_us(unsigned int us) {
  *
  * This method gets the current-timestamp in micro-seconds.
  */
-unsigned long
+unsigned long long
 pelion_get_current_timestamp_us() {
 
     return 0;
